Update services
Set up services as singletons so that they can be called for by name
across models and aborted or called. This new architecture affords the
developer the ability to control services in progress and negates the
unnecessary prior ability to instantiate multiple services of the same
origin on the fly.